Fix clean build on JDK8
TREC recently moved their website to HTTPS, but it seems the certificate is (at least) not trusted on JDK8 on macOS.

Instead of forcing people to update the Java certificate store manually, remove the automatic download of the trec_eval tool.

Once JDK8 is EOL in September 2018, we can revert this.