Update Lexigram response format to new format
It seems Lexigram provides a different JSON format when running on Travis CI.

Instead of returning all results in the `array` key, it returns ``"array": false` with results ONLY in the `object` key.

As locally all results are provided in both keys, fallback to the `object` key as default, so that it works every time.