Repaired tests concerning property injection into templates.
I had refactored existing tests to use the StaticMapQueryDecorator
which in the first try wouldn't work. It is, unfortunately, important
to keep the StaticMapQueryDecorator within a TemplateDecorator.
If the TemplateDecorator is the delegate decorator for the
StaticMapQueryDecorator, the original template won't be loaded because
the StaticMapQueryDecorator expects the template to already be there
when calling "query(Topic)", eventually resulting in an NPE.