Package refactoring, addition of classes for LtR feature management.
Refactoring: Since new development for this fork is oved to the
JULIELab it seems right to put the new developments under the package
de.julielab. Because the PM classifier was also done at the JULIE Lab,
I moved it into de.julielab.

LtR feature management: I want to be able to toggle features and
feature groups for experiments on which features/feature groups
have what impact on NDCG peformance. Nothing of this does actually
work yet, this is an early implementation stage.