Added first features and a possibility to retrieve full documents.
The TfidfFeatureGroup is basically a copy from the code in the PM
classifier TfidfPipe. It won't work right now because the
secondstring TFIDF has to be trained first, but in principle the code
should do.
For LtR feature generation we need the full document text of the
documents to be ranked. Since I use UIMA for processing I here
stick to the UIMA format and thus also to the document database
I employ. Thus, the OriginalDocumentRetrieval class returns
UIMA CAS objects retrieved from the database.
The TrecConfig object has been extended for configuration parameters
required by the new classes.