Added a working test for original document retrieval.
The test assumes that the config/costosys.xml file points to
a valid postgres database. It currently points to "localhost" which
reflects the fact that I'm using an SSH tunnel to our internal Postgres
database. Of course the test will only work with a filled database
using the UIMA components in the `uima` directory.