Adding a CAS pool and polishing CAS handling.
We will need to create CAS objects on request because CASes
are rather expensive to create and to hold in memory. For example,
I was not able to create 10000 CAS instances on my Laptop because
the memory (I could give 6GB) was not enough. To keep flexible
and swift, we use a CasPool. This will of course make some coding
a bit more complicated because the CAS objects have to be filled
with data, processed and released.