Logic-retaining refactoring of the ElasticSearchQuery.
I would like a more flexibel query model where the information
on what index to query is largely taken from the Topics themselves
because this facilitates cross-dataset experiments.

The TopicSet internally now uses a List actually. This is because
the TopicSetTest expected the topics to exist in order which was
broken when I added some fields to the topics (HashSets do not have
a reliable order). A list seems to do just fine.

All non-ignored test working, except Lexigram which wasn't run.