Added a configuration setting to make the document cache read-only.
This is the best solution I could find to allow multiple processes
access the document cache simultaneously. Once the cache is completely
written, there is no difference anyway because no information is missing.
The sad thing is that one has now always to remember to build the
cache once after a fresh checkout.