Added a cache for ElasticSearch results.
The cache keys are the complete JSON query strings sent to ES.
Thus, as long as the index contents do not change - which they
shouldn't for a given task while evaluating - the result will
always be the same for the same query, making it a perfect fit
for caching. Especially since some queries (phrase, for example) take
quite some time to run.