Fixed issues with multiple use of the same Query object.
The DynamicQueryDecorator added gene and disease synonyms each time
`query()` was called, resulting in longer and longer lists of the
very same query terms.
Also, since for each Query a new `ElasticSearch` instance was
created, caching was flawed regarding the readOnly flag.

The DynamicQueryDecorator now keeps track of the topics that it
has already expanded. Added the CacheService to solve the caching issues.