Working on UIMA issues. Updated the gradle wrapper in UIMA components.
The "UIMA issues" is actually just the situation that some documents
do not seem to have gotten a PM classification in my database,
namely 150 documents are missing. Also, I would like to access the
classification result from the GRU classification which is currently
only written into the index directly without storing it to the
database which I want to fix now.