Added an ES mapping template and a class to create ES indexes with those.
For the LtR experiments, we need different IR measures as features
for the LtR algorithm. In ES, we need different indexes for that
because the similarity is set on a field basis.
But the parameters of a similarity can be changed on an existing index,
making experiments with different similarity parameters possible.