Created SimilarityParameter classes for changing similarities on the fly.
Each possible similarity has one parameter settings class now. Those
can be given to the ElasticSearchQuery which in turn forwards them
to the ElasticSearch instance created on `query(Topic)`. The
ElasticSearch instance will then reconfigure the index for the
parameters set.
It is important to keep in mind that this prohibits multiple
simultaneous experiments on the same index! This is due to the fact
that the similarity parameter settings are set per index (not per
 query!) and also that an index must be closed for changing the parameters.