All tests running when using the remote cache.
Whether to use remote or local is set in the
`trec-pm.properties` file for the property `cache.type`. It can
have the values `local` or `remote`. When `remote` is chosen,
the `cache.host` and `cache.port` properties must point to the
that has been separately started with these settings.
For this, the `scripts/startRemoteCache.sh` script can be used after
having `mvn package`'d the project.