Added the `FeatureUtils` class. It currently offers string normalization.
This is for LtR feature generation.
This will be used to find the topic disease and gene names in the
document text. The normalization should help for finding minor variants.

Also added the `de.julielab.jcore.types.PointMutation` to the list of
annotation types to be loaded. This will be used to check if the
topic's gene variant is mentioned in the document.