Subset documents only when writing
We need the full set of documents (including unjudged) in memory so that we don't add new documents to be judged that are already in the list.

So, load all documents while reading and subset unjudged documents when writing a qrel file. This actually make the code more robust, because a qrel file MUST NOT contain negative relevance assessments, which could originally be added by mistake.