Make `Topic` do not depend on `Task`
The only technical reason to have a `Topic` depend on `Task` was to be able to reuse `Retrieval` and `ElasticSearchQuery` objects. However, they should not be reused for different tasks because it's hard to keep state with a decorator chain.

This removes such dependence to make code simpler. It also removes an `ElasticSearchQuery` constructor that allowed such reuse (with an unused `GoldStandard` parameter).