Resolves #79.
However, the current construct is fragile. The reranking score must be
set to the results that end up being written. The score is saved
in the Documents returned from the ranker. However, the ranker might
add a rather arbitrary `IRScore` type. This should be synchronized,
e.g. by retrieving the type of `IRScore` set by the ranker.