Moved test templates und subtemplates to new directories. Adapted tests.
I wanted to create a test checking the current results in our TREC19 runs so that we can see quickly when performance changes unintentionally.
However, being a test, the `src/test/resources/subtemplates` was used automatically by the classpath lookup mechanism rather then the desired `src/main/resources/subtemplates` directory. To fix this, the `templates` and `subtemplates` directories in `src/test/resources` were renamed to `test-templates` and `test-subtemplates`, respectively. The tests are now adapted to the change and work fine.