Going towards parameter optimization for booster keywords. Currently: ConditionalCancer.
The new `FeatureControlledConditionalCancerQueryDecorator` works just like `ConditionalCancerQueryDecorator` with the exception that the actual keywords come from the feature configuration. For this purpose, the names of elements at specific configuration paths are directly used as keywords. Thus, the feature configuration does not just determine which keywords are active but it also defines the available keywords themselves. This is a good thing because otherwise we would need to check each possible keywords in the code for its (de-)activation status. Now we just check if some element at some path has the value `true` and, if so, we add the element name as a keyword.
This will also be done with other keywords lists.