Identified - not really fixed - LtR feature creation issue regarding query topic expansions.
The issue is that the topic dependent features are created on the topic objects as they come. There is no control or check about topic expansion or not. Thus, it is perfectly possible right now to use some query expansion at train time and not use the same expansion at test time. While this could easily be considered an inconsistency error on part of the programmer, this is a detail easy to miss (but with large consequences since topic expansion is a major mechanic) and should be checked and reported somehow.