Identified and fixed the 'issue' with a better 2019PM result due to double application of synonyms.
I had added a set to the `DynamicQueryDecorator` that kept track of the topics already expanded. This did work as intended, however, since for each now experiment a new retrieval with new decorators is created. Thus, the new decorators have no knowledge of already decorated topics. But since the very same topic instances are given to the decorators, there is the multiapplication.

For now this has been fixed be setting a field into the `QueryDescription` class keeping track of the decorators that have been applied to the query. This information is used in the `DynamicQueryDecorator` to avoid multiple application of the same decorator to the same query.

However, this does not solve the issue that the topics are stateful objects that "drag" information from one experiment to another. This should be remedied by always beginning with a clean copy of the topics, given by the respective method `getCleanCopy()`. I just don't know where to apply this yet.