Fixed tests. TopicMatch feature generation has been overhauled.
The FeatureControlCenter had to be reset to the prerequisites of the tests, the CacheService had to be initialized in more tests.
The TopicMatch feature generation classes now work with clean copies of the respective topics. They also avoid duplicates of disease synonyms/hypernyms and reject synonyms/hypernyms equaling the topic disease term.