Finished first untested version of CT parameter optimization templates. Added a new EvaluationRoute to the Http server for the clinical trials and added all the official datasets as instance names to the EvaluationRoute.
The last point means that we can specify an instance like `ct2018` or `pm2017` to evaluate the respective dataset with the given parameters. Note that `ct2017` wont work because we currently evaluate for `infNDCG` and we can't copute that for `ct2017` as no sample gold standard was created back then.