The update of the index settings for the similarity parameters is now working fine.
We now check if there are shards with red status. If so, we wait 200ms and check again until there are no red shards left. Then we continue. This seems to work fine.
Also fixed a bug where the current and desired parameter value difference was not checked neutral to the sign, needed Math.abs().